Not entirely awful solution
Initially I passed the test by just "puts" the entire thing. Worked, but
not extensible.

This way is...better? I don't think what it's doing is opaque, but I'm
not sure the extra object really flushes out the concepts I'm trying to
get at.

For example, the stanzas are really made up of the structure: "[noun] that [verb]".
But I'm ignoring that for now and just treating them as whole strings.
THat means adding a new stanza requires someone to come in and edit the
Stanza collection. No onerous, but error prone.

A more open/close solution would maybe make a more abstract Stanza class
and a collection of Stanzas that could be edited more easily. I'm
actually attracted to YAML as a solution here. The stanza data could
just be configuriation details.