Extraction of a Stanza
Here we extract out the stanza logic from the Stanzas. This is better.

I still find this hard to extend. A requirement change for a new stanza
or a reordering of stanzas would require opening up existing code. But
absent those requirements, I'm not sure there's much more to do here.